test: reduce code duplication after splitting files in cloud/aws-sdk/v3/dynamodb
